requirements.txt
google-adk>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
google-generativeai>=0.3.0
pathlib>=1.0.1
fastmcp>=0.1.0
uvicorn>=0.15.0 